feat: use new compiler and new decorators
This update removes the need for a separate decorators package
and updates the support for ES decorators to the current stage-3
proposal. Currently private methods/fields are not supported by the
compiler, but support can be added in the future.
In addition, reflection data for classes are now provided by the
compiler itself and not derived at run-time.